+++ config/bundles.php
@@ -7,4 +7,5 @@ return [
     Symfony\Bundle\WebProfilerBundle\WebProfilerBundle::class => ['dev' => true, 'test' => true],
     Symfony\Bundle\MakerBundle\MakerBundle::class => ['dev' => true],
     League\FlysystemBundle\FlysystemBundle::class => ['all' => true],
+    EmberNexusBundle\EmberNexusBundle::class => ['all' => true],
 ];