+++ lib/EmberNexusBundle/src/EmberNexusBundle.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace EmberNexusBundle;
+
+use Symfony\Component\HttpKernel\Bundle\Bundle;
+
+class EmberNexusBundle extends Bundle
+{
+}